prerelease 1.9 R0

+++ components/xpt2046/xpt2046.c
@@ -66,7 +66,7 @@ void xpt_init()
 	gpioConfig.mode         = GPIO_MODE_OUTPUT;
 	gpioConfig.pull_up_en   = GPIO_PULLUP_ENABLE;
 	gpioConfig.intr_type    = GPIO_INTR_DISABLE;
-	gpioConfig.pin_bit_mask = (1<<csPin);
+	gpioConfig.pin_bit_mask = ((uint64_t)(((uint64_t)1)<<csPin));
 	ESP_ERROR_CHECK(gpio_config(&gpioConfig));		
 	gpio_set_level(csPin, 1);
 	ESP_LOGI(TAG, "Touch cs: %d init done",csPin);
